requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
PyJWT>=2.8.0
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0
